pyproject.toml
[project]
name = "tabrikchi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aerich>=0.9.2",
    "aiogram>=3.22.0",
    "anyio>=4.11.0",
    "python-decouple>=3.8",
    "tomlkit>=0.13.3",
    "tortoise-orm[asyncpg]>=0.25.1",
]
